requirements.txt
numpy
torch
pillow
scikit-learn
matplotlib
seaborn
pandas
tqdm
ftfy
regex
requests
litellm
git+https://github.com/openai/CLIP.git

# Optional: for Image Retrieval (nimitz retrieve ...)
# Set environment variables:
# - UNSPLASH_ACCESS_KEY for Unsplash (get at https://unsplash.com/developers)
# - PEXELS_API_KEY for Pexels (get at https://www.pexels.com/api/)

# LLM Support - now REQUIRED for smart entity filtering
# Supports: OpenAI (GPT-4V), Anthropic (Claude), Google (Gemini)
# Set at least one: ANTHROPIC_API_KEY, GEMINI_API_KEY, or OPENAI_API_KEY

# Optional: for PDF export (nimitz export-pdf)
# pip install reportlab

# Development/Testing
# pip install pytest